Always prefetch fixed amount
Prior to the change the prefetch amount was coupled with size; it
would prefetch the size requested. This was changed to use a fixed
amount to have similar behavior as the standard library io module,
where prefetch is a separate configuration from size. In the future,
we can make the prefetch amount configurable.