Regenerate requirements-dev.txt with uv
We were currently using pip-tools to compile the lockfile however
as we have been adding dependencies there have been branching of the
dependency tree due to Python version and platform. Because pip-tools
does not proxy the environment markers, it forces us to be very specific on
the environment where we compile the lock file (right now is Windows,
Python 3.9) to have it work on all of the environments we support.

With uv, it is an new Python package manager that has the ability to
export a requirements lock file in the same style as pip-tools. More
importantly, it handles environment markers meaning we can generate
the lock file on any environment and have the lock file work for any
enviroment.

This is important because in adding torch as a dependency it pulls in
several dependencies just for Linux. In order to capture these branching
dependencies, it would require compiling a requirements lock file on
both Windows and Linux. So by utilizing uv for the lock file, we are
able to avoid this extra overhead.