Add support for map-style datasets
This dataset class is intended for random sampling of dataset loaded
into memory. The dataset class itself allows for accessing blobs listed
from a container (e.g., from_container_url) or a customer provied list of
blobs (e.g., from_blob_urls). In addition, customers can provide a
custom transform that accepts a "Blob" object that can be used to
access blob content and identifiers as part of the transform.

One additional refactor that needed to be made was to make executor
instantiation in the blob client lazy. The executor itself is not
pickleable when the PyTorch data loader uses workers. By making it
lazy, it ensures that it is created once the process have been
created and start reading data