Add support for map-style datasets
This dataset class is intended for random sampling of dataset loaded
into memory. The dataset class itself allows for accessing blobs listed
from a container (e.g., from_container_url) or a customer provied list of
blobs (e.g., from_blob_urls). In addition, customers can provide a
custom transform that accepts a "Blob" object that can be used to
access blob content and identifiers as part of the transform.