Introduce AzStorageBlobClientFactory (#27)
The abstraction consolidates AzStorageBlobClient creation allowing:

1. Any SDK specific resolution (e.g., credential, default client
   configurations) to be consolidated to one class instead of being
   split between multiple classes
2. Enables better resource sharing across clients. Currently, it just
   shares the transport and credential object. In later commits, we
   will explore sharing credential caching and/or pipelines

Both of these properties will be especially useful for dataset
implementation as it will live in new classes and may create more than
one client.

In addition, the commit updates the typing of the io module to access
_client types as a property of the _client module. This made imports
less verbose/easier to follow when trying to prevent non-public classes
from leaking into io module which is public.