Add echo client request policy
Policy was added to ensure that requests made by the client match
responses returned. This is an extra guard rail to ensure when using
an azstoragetorch interface with processes, customers will not
successfully process a response from a different request because
the underlying connection resource was shared as part of forking.

This was seen as part of the BlobDataset.from_container_url() where
it would:
1. Call List blobs on instantiation in the main process
2. Make API calls in the workers to retrieve blob properties and content
3. Receive modified access condition exceptions because the blob
   properties returned to a client in the worker was not the correct
   ETag.

While the BlobDataset was made to defer listing blobs to workers to
prevent using requests session in the main process. This edge case
can still be ran into if accessing the dataset directly in the main
process and then passing it to a data loader with workers set.