Add echo client request policy
Policy was added to ensure that requests made by the client match
responses returned. This is an extra guard rail to ensure when using
an azstoragetorch interface with processes, customers will not
successfully process a response from a different request because
the underlying connection resource was shared as part of forking.

This was seen as part of the BlobDataset.from_container_url() where
it would:
1. Call List blobs on instantiation in the main process
2. Make API calls in the workers to retrieve blob properties and content
3. Receive modified access condition exceptions because the blob
   properties returned to a client in the worker was not the correct
   ETag.

While the BlobDataset is going to be set up in a later commit to
not share transports between container and blob clients. This edge case
can still be ran into if accessing the dataset directly in the main
process and then passing it to a data loader with workers set.

A custom exception class was created to avoid using an AzureError
based exceptions which would be retried. We chose not to retry here
to avoid sending unnecessary requests stemming from not recommended
dataset usage patterns (i.e., access dataset in the main process
and then provide it to a multi-worker data loader)